Make switch statements work in new Clang
The -Wno-switch-default and -Wcovered-switch-default make it
impossible to write a warning-free switch statement covering all
the values of an enum. Since -Wcovered-switch-default lets us have
a warning if we add an enum value but forget to update all the
switch statements that are affected (except those that already did
not cover all the cases individually), that warning is better.

The -Wno-switch-default warning appears to be new in Clang 18.